Remove empty nodes when deleting cookies.
BUG=39389
TEST=none

Review URL: http://codereview.chromium.org/1751013

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@45575 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 241a4fe1a3ca00f365bab107aa5736eb27ed222d

+++ chrome/browser/cookies_tree_model.cc
@@ -409,11 +409,15 @@ void CookiesTreeModel::DeleteAllStoredObjects() {
 }
 
 void CookiesTreeModel::DeleteCookieNode(CookieTreeNode* cookie_node) {
+  if (cookie_node == GetRoot())
+    return;
   cookie_node->DeleteStoredObjects();
   // find the parent and index
   CookieTreeNode* parent_node = cookie_node->GetParent();
   int cookie_node_index = parent_node->IndexOfChild(cookie_node);
   delete Remove(parent_node, cookie_node_index);
+  if (parent_node->GetChildCount() == 0)
+    DeleteCookieNode(parent_node);
 }
 
 void CookiesTreeModel::UpdateSearchResults(const std::wstring& filter) {
